Adding a GraphQL endpoint.
This GraphQL endpoint uses graphql-java (https://www.graphql-java.com/) and the
the graphql spring boot library to provide an HTTP endpoint for GET and POST
methods.  The spring boot library generates an endpoint at /graphql.

This endpoint is wired into the existing search service to provide the same
functionality is the REST endpoint /search.